requirements.txt
ccxt
openai
pandas
python-dotenv
requests
